requirements.txt
# pip install -r requirements.txt
Django
djangorestframework
pytz
Pillow
requests
